refactor: move add_properties into a maintain module directory
`mutation/maintain.rs` reached 2522 lines, past the 2500-line
source-quality ceiling. `add_properties` is the separable half: unlike
every other operation in the file it writes no topology, and its private
vocabulary (aggregate/spatial expression parsing, the ancestor walk,
geometry resolution) has no other caller.

`maintain.rs` becomes `maintain/mod.rs` (1946 lines) plus
`maintain/add_properties.rs` (589). The flat `maintain_edge_spec_tests.rs`
moves in as `maintain/edge_spec_tests.rs`, dropping its `#[path]`
attribute. `AddPropertiesReport` was never reachable through a public
path, so the new module stays `pub(crate)` and only the two items
`api::mutation` re-exports are lifted back into `maintain`.

The four source-quality function exceptions are re-pathed with their
recorded metrics untouched — nothing widened.
